+++ R/DiSCo.R
@@ -28,9 +28,9 @@ utils::globalVariables(c("y_col", "id_col", "time_col", "t_col", "group", "x", "
 #' @param q_max Numeric, maximum quantile to use. Set this together with `q_min` to restrict the range of quantiles used to construct the synthetic control.
 #' Default is 1 (all quantiles). Currently NOT implemented for the `mixture` approach.
 #' @param CI Logical, indicating whether to compute confidence intervals for the counterfactual quantiles. Default is FALSE.
-#' @param CI_placebo Logical, indicating whether to compute confidence intervals for the pre-treatment periods. Default is TRUE.
-#' If you have a lot of pre-treatment periods, setting this to FALSE can speed up the computation.
+#' The confidence intervals are computed using the bootstrap procedure described in \insertCite{vandijcke2024rto;textual}{DiSCos}.
 #' @param boots Integer, number of bootstrap samples to use for computing confidence intervals. Default is 500.
+#' @param replace Logical, indicating whether to sample with replacement when computing the bootstrap samples. Default is TRUE.
 #' @param cl Numeric, confidence level for the (two-sided) confidence intervals.
 #' @param graph Logical, indicating whether to plot the permutation graph as in Figure 3 of the paper. Default is FALSE.
 #' @param qmethod Character, indicating the method to use for computing the quantiles of the target distribution. The default is NULL, which uses the \code{\link[stats]{quantile}} function from the stats package.
@@ -45,19 +45,39 @@ utils::globalVariables(c("y_col", "id_col", "time_col", "t_col", "group", "x", "
 #' also provide a list of support points in the `grid.cat` parameter. When that is provided, this parameter is automatically set to TRUE. Default is FALSE.
 #' @param grid.cat List, containing the discrete support points for a discrete grid to be used with the mixture of distributions approach.
 #' This is useful for constructing synthetic distributions for categorical variables. Default is NULL, which uses a continuous grid based on the other parameters.
-#' @return A list containing, for each time period, the elements described in the return argument of \code{\link{DiSCo_iter}}, as well as the following additional elements:
+#' @return A list containing the following elements:
+#' \itemize{
+#' \item \code{results.periods} A list containing, for each time period, the elements described in the return argument of \code{\link{DiSCo_iter}}, as well as the following additional elements:
 #' \itemize{
 #'  \item \code{DiSco}
 #'  \itemize{
-#'  \item \code{CI } A list with the confidence intervals and standard errors for the counterfactual quantiles, if `CI` is TRUE and for the periods specified in `CI_periods`.
-#'  See the output of \code{\link{DiSCo_CI}} for details.
 #'  \item \code{quantile } The counterfactual quantiles for the target unit.
 #'  \item \code{weights } The optimal weights for the target unit.
 #'  \item \code{cdf } The counterfactual CDF for the target unit.
 #'  }
-#'
+#'  }
+#' \item \code{weights} A numeric vector containing the synthetic control weights for the control units, averaged over time.
+#' When `mixture` is TRUE, these are the weights for the mixture of distributions, otherwise they are the weights for the quantile-based approach.
+#' \item \code{CI} A list containing the confidence intervals for the counterfactual quantiles and CDFs, if `CI` is TRUE.
+#' Each element contains two named subelements called `upper`, `lower`, `se` which
+#' are the upper and lower confidence bands and the standard error of the estimate, respectively.
+#' They are G x T matrices where G is the specified number of grid points and T is the number of time periods.
+#' The elements are:
+#' \itemize{
+#' \item \code{cdf} The bootstrapped CDF
+#' \item \code{quantile} The bootstrapped quantile
+#' \item \code{quantile_diff} The bootstrapped quantile difference
+#' \item \code{cdf_diff} The bootstrapped CDF difference
+#' \item \code{bootmat} A list containing the raw bootstrapped samples for the counterfactual quantiles and CDFs, if `CI` is TRUE.
+#' These are not meant to be accessed directly, but are used by `DiSCoTEA` to compute aggregated standard errors. Advanced users
+#' may wish to access these directly for further analysis. The element names should be self-explanatory.
+#' #' \item \code{control_ids} A list containing the control unit IDs used for each time period, which can be used to identify the weights
+#' associated with each control as the returned weights have the same order as the control IDs.
 #' \item \code{perm } A \code{\link{permut}} object containing the results of the permutation method, if `permutation` is TRUE.
 #' Call `summary` on this object to print the overall results of the permutation test.
+#' #' \item \code{evgrid} A numeric vector containing the grid points on which the quantiles were evaluated.
+#' \item \code{params} A list containing the parameters used in the function call.
+#' }
 #' }
 #' @importFrom Rdpack reprompt
 #' @importFrom stats sd quantile rnorm
@@ -65,10 +85,9 @@ utils::globalVariables(c("y_col", "id_col", "time_col", "t_col", "group", "x", "
 #' @references
 #'  \insertAllCited()
 #' @export
-
-
+#'
 DiSCo <- function(df, id_col.target, t0, M = 1000, G = 1000, num.cores = 1, permutation = FALSE, q_min = 0, q_max = 1,
-                  CI = FALSE, CI_placebo=TRUE, boots = 500, cl = 0.95, graph = FALSE,
+                  CI = FALSE, boots = 500, replace=TRUE, cl = 0.95, graph = FALSE,
                   qmethod=NULL, seed=NULL, simplex=FALSE, mixture=FALSE, grid.cat=NULL) {
 
   #---------------------------------------------------------------------------
@@ -79,7 +98,7 @@ DiSCo <- function(df, id_col.target, t0, M = 1000, G = 1000, num.cores = 1, perm
 
   # check the inputs
   checks(df, id_col.target, t0, M, G, num.cores, permutation, q_min, q_max,
-         CI, CI_placebo, boots, cl, graph,
+         CI, boots, cl, graph, # TODO: update this
          qmethod, seed)
 
   if (!is.null(grid.cat)) { # TODO: allow restricted grid for mixture
@@ -133,13 +152,11 @@ DiSCo <- function(df, id_col.target, t0, M = 1000, G = 1000, num.cores = 1, perm
   #---------------------------------------------------------------------------
   ### calculate average weights across pre-periods
   #---------------------------------------------------------------------------
-  Weights_DiSCo_avg <- results.periods$`1`$DiSCo$weights
-  Weights_mixture_avg <- results.periods$`1`$mixture$weights
-  if (T0 > 1) {
-    for (yy in 2:T0){
+  Weights_DiSCo_avg <- 0
+  Weights_mixture_avg <- 0
+  for (yy in 1:T0){
       Weights_DiSCo_avg <- Weights_DiSCo_avg + results.periods[[yy]]$DiSCo$weights
       Weights_mixture_avg <- Weights_mixture_avg + results.periods[[yy]]$mixture$weights
-    }
   }
   Weights_DiSCo_avg <- (1/T0) * Weights_DiSCo_avg
   Weights_mixture_avg <- (1/T0) * Weights_mixture_avg
@@ -153,6 +170,7 @@ DiSCo <- function(df, id_col.target, t0, M = 1000, G = 1000, num.cores = 1, perm
   } else {
    cdf <- lapply(seq(1:T_max), function(x) results.periods[[x]]$controls$cdf[,-1] %*% as.vector(Weights_mixture_avg) )
   }
+
   # calculating the counterfactual target quantiles and CDF
   for (x in seq(1:T_max)) {
     if (!mixture) { # if DiSCo
@@ -178,30 +196,32 @@ DiSCo <- function(df, id_col.target, t0, M = 1000, G = 1000, num.cores = 1, perm
     weights <- Weights_DiSCo_avg
   }
 
-  # calculate confidence intervals for selected time periods
-  if (CI & CI_placebo) { # if wants CI for all periods
-    CI_periods <- seq(1, T_max)
-  } else if (CI & !CI_placebo) {
-    CI_periods <- seq(T0+1, T_max)
-  } else{
-    CI_periods <- seq(1, T_max)
-  }
+  #---------------------------------------------------------------------------
+  ### confidence intervals
+  #---------------------------------------------------------------------------
+
+
   if (!is.null(grid.cat)) grid <- grid.cat
-  for (x in CI_periods) {
-    if (CI) cat(paste0("Computing confidence intervals for period: ", x + t_min - 1, "\n"))
-    controls <- results.periods[[x]]$controls$data
-    if (is.null(grid.cat)) grid <- results.periods[[x]]$target$grid
-    if (CI) {
-      CI_temp <- DiSCo_CI(controls=controls, weights=weights, grid=grid,
-                          mc.cores=num.cores, cl=cl, num.redraws=boots, evgrid = evgrid,
-                          qmethod=qmethod, mixture=mixture)
-    } else {
-      CI_temp <- NULL
-    }
-    results.periods[[x]]$DiSCo$CI <- CI_temp
+  if (CI) {
+    controls <- lapply(results.periods, function(x) x$controls$data)
+    target <- lapply(results.periods, function(x) x$target$data)
+    grid <- lapply(results.periods, function(x) x$target$grid)
+
+    ## bootstrap the estimator
+    CI_bootmat <-  mclapply.hack(1:boots, DiSCo_CI, controls=controls,target=target,
+                              T_max=T_max, T0=T0, grid=grid, mc.cores=num.cores,
+                              evgrid=evgrid, mixture=mixture, M=M, simplex=simplex,
+                              qmethod=qmethod, replace=replace)
+    CI_out <- parseBoots(CI_bootmat, cl)
+
+  } else {
+    CI_out <- NULL
   }
 
-  # permutation tests
+
+  #---------------------------------------------------------------------------
+  ### permutation test
+  #---------------------------------------------------------------------------
   if (permutation) {
 
     # run the permutation test
@@ -220,9 +240,18 @@ DiSCo <- function(df, id_col.target, t0, M = 1000, G = 1000, num.cores = 1, perm
   # rename periods for user convenience
   names(results.periods) <- t_min  + seq(1:T_max) - 1
 
+  if (mixture) {
+    weights <- Weights_mixture_avg
+  } else {
+    weights <- Weights_DiSCo_avg
+  }
 
-  return(list(results.periods=results.periods, Weights_DiSCo_avg=Weights_DiSCo_avg,
-              Weights_mixture_avg=Weights_mixture_avg, control_ids=controls.id, perm=perm_obj, evgrid=evgrid, params=list(df=df_pres, id_col.target=id_col.target,
-                                                                                                                          t0=t0, M=M, G=G, CI=CI, cl=cl, CI_placebo=CI_placebo, qmethod=qmethod, boot=boots, q_min=q_min, q_max=q_max)))
+  #---------------------------------------------------------------------------
+  return(list(results.periods=results.periods, weights=weights, CI=CI_out,
+              control_ids=controls.id, perm=perm_obj, evgrid=evgrid,
+              params=list(df=df_pres, id_col.target=id_col.target,
+              t0=t0, M=M, G=G, CI=CI, cl=cl, qmethod=qmethod,
+              boot=boots, q_min=q_min, q_max=q_max))
+         )
 
 }