+++ R/DiSCo_mixture.R
@@ -16,7 +16,8 @@
 #' \item \code{weights.opt } The optimal weights.
 #' }
 #' @keywords internal
-DiSCo_mixture <- function(controls1, target, grid.min, grid.max, grid.rand, M, simplex) {
+DiSCo_mixture <- function(controls1, target, grid.min, grid.max, grid.rand, M,
+                          simplex) {
 
   ###### The mixture of distributions approach
   # we again only focus on the first half of the data
@@ -37,9 +38,34 @@ DiSCo_mixture <- function(controls1, target, grid.min, grid.max, grid.rand, M, s
     CDF.matrix[,(ii+1)] <- CDF.control[[ii]](grid.rand)
   }
 
+
+  # Solving the convex problem with CVXR
+  out <- DiSCo_mixture_solve(length(controls1), CDF.matrix,  grid.min, grid.max, grid.rand, M, simplex)
+  out[['cdf']] <- CDF.matrix
+
+  return(out)
+}
+
+
+#' @title DiSCo_mixture_solve
+#' @description The solver for the alternative mixture of distributions approach in the paper
+#' @param c_len The number of controls
+#' @param CDF.matrix The matrix of CDFs
+#' @inheritParams DiSCo_mixture
+#' @return A list containing the following elements:
+#' \itemize{
+#' \item \code{distance.opt } The optimal value of the Wasserstein distance.
+#' \item \code{mean } The optimal value of the Wasserstein barycenter.
+#' \item \code{target.order } The target unit, ordered.
+#' \item \code{weights.opt } The optimal weights.
+#' }
+#' @keywords internal
+DiSCo_mixture_solve <- function(c_len, CDF.matrix, grid.min, grid.max, grid.rand,
+                                M, simplex) {
+
   # Solving the convex problem with CVXR
   # the variable we are after
-  theweights <- CVXR::Variable(length(controls1))
+  theweights <- CVXR::Variable(c_len)
   # the objective function
   objective <- CVXR::cvxr_norm((CDF.matrix[,2:ncol(CDF.matrix)] %*% theweights - CDF.matrix[,1]))
 
@@ -66,8 +92,7 @@ DiSCo_mixture <- function(controls1, target, grid.min, grid.max, grid.rand, M, s
   themean.order <- themean[order(grid.rand, decreasing=FALSE)]
   target.order <- CDF.matrix[order(grid.rand, decreasing=FALSE),1]
 
-
   return(list("weights.opt" = theweights.opt, "distance.opt" = thedistance.opt,
-              "mean" = themean.order, "target.order" = target.order, "cdf" = CDF.matrix))
+        "mean" = themean.order, "target.order" = target.order))
 
 }